// mobile/lib/types.ts
export const ClientSchema = z.object({
    name: z.string().min(3, { message: "O nome deve ter pelo menos 3 caracteres." }),
    phone: z.string().min(10, { message: "O telefone deve ter pelo menos 10 caracteres." }),
    cpfCnpj: z.string().optional(),
    email: z.string().email({ message: "Formato de e-mail inválido." }).optional().or(z.literal('')),
    address: z.string().min(5, { message: "O endereço deve ter pelo menos 5 caracteres." }),
    latitude: z.preprocess(toNumOrNull, z.number().min(-90).max(90).nullable()).optional(),
    longitude: z.preprocess(toNumOrNull, z.number().min(-180).max(180).nullable()).optional(),
    googleMapsLink: z.string().url({ message: "Por favor, insira um link válido do Google Maps." }).optional().or(z.literal('')),
    observations: z.string().optional(),
});
// ...
export type Client = z.infer<typeof ClientSchema> & { id: string, createdAt: any };